requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
axial-positional-embedding=0.2.1=pypi_0
blas=1.0=mkl
brotli-python=1.0.9=py38h6a678d5_8
bunch=1.0.1=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.3.11=h06a4308_0
certifi=2024.7.4=py38h06a4308_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.6=pypi_0
contourpy=1.1.1=pypi_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.12.1=pypi_0
deprecated=1.2.14=pypi_0
einops=0.4.0=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.15.4=pypi_0
fonttools=4.52.4=pypi_0
freetype=2.12.1=h4a9f257_0
fsspec=2024.6.1=pypi_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
huggingface-hub=0.23.4=pypi_0
idna=3.7=py38h06a4308_0
importlib-metadata=7.1.0=pypi_0
importlib-resources=6.4.0=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
joblib=1.4.2=pypi_0
jpeg=9e=h5eee18b_1
kiwisolver=1.4.5=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libwebp-base=1.3.2=h5eee18b_0
llvmlite=0.41.1=pypi_0
local-attention=1.4.4=pypi_0
lz4-c=1.9.4=h6a678d5_1
matplotlib=3.7.0=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.8=py38h5eee18b_0
mkl_random=1.2.4=py38hdb19cb5_0
mpmath=1.3.0=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
numba=0.58.1=pypi_0
numpy=1.23.5=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h9ca470c_1
openssl=3.0.14=h5eee18b_0
packaging=24.0=pypi_0
pandas=1.5.3=pypi_0
patool=1.12=pypi_0
pillow=10.3.0=py38h5eee18b_0
pip=24.0=py38h06a4308_0
product-key-memory=0.1.10=pypi_0
pyparsing=3.1.2=pypi_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.19=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
pytorch=1.12.1=py3.8_cuda11.3_cudnn8.3.2_0
pytorch-mutex=1.0=cuda
pytorch-ranger=0.1.1=pypi_0
pytz=2024.1=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
reformer-pytorch=1.4.4=pypi_0
requests=2.32.2=py38h06a4308_0
safetensors=0.4.3=pypi_0
scikit-learn=1.2.2=pypi_0
scipy=1.10.1=pypi_0
setuptools=69.5.1=py38h06a4308_0
six=1.16.0=pypi_0
sktime=0.16.1=pypi_0
sqlite=3.45.3=h5eee18b_0
sympy=1.11.1=pypi_0
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=3.5.0=pypi_0
timm=1.0.7=pypi_0
tk=8.6.14=h39e8969_0
torch-optimizer=0.3.0=pypi_0
torchaudio=0.12.1=py38_cu113
torchdiffeq=0.2.4=pypi_0
torchvision=0.13.1=py38_cu113
tqdm=4.64.1=pypi_0
typing-extensions=4.12.0=pypi_0
urllib3=2.2.2=py38h06a4308_0
wheel=0.43.0=py38h06a4308_0
wrapt=1.16.0=pypi_0
xz=5.4.6=h5eee18b_1
zipp=3.19.0=pypi_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.5=hc292b87_2
